@@ -3,17 +3,18 @@
 package com.mti.todo_app_with_firebase.UI.Main;
 
 import android.app.SearchManager;
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProviders;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.SearchView;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -96,7 +97,7 @@ public class MainActivity_Recycler_Fragment extends Fragment implements MainActi
         super.onActivityCreated(savedInstanceState);
 
         //initializing viewmodel
-        mMain_recyclerFragment_viewModel=ViewModelProviders.of(this).get(Main_RecyclerFragment_ViewModel.class);
+        mMain_recyclerFragment_viewModel= ViewModelProviders.of(this).get(Main_RecyclerFragment_ViewModel.class);
 
 
         mRecyclerView = getView().findViewById(R.id.listTodo_fragment);
