@@ -6,10 +6,10 @@
 
 package com.mti.todo_app_with_firebase.Repository.Main.Room;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
 
 import com.mti.todo_app_with_firebase.model.ToDo;
 
